generalized start commands
